+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 Unreleased
 ----------
 
+* glean_parser now supports Python versions 3.5, 3.6, 3.7 and 3.8.
+
 1.13.0 (2019-12-04)
 -------------------